+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add `LivestreamChatState.typingMemberInfos` for accessing channel-member info of typing users [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 - Add `LivestreamChatState.typingUsersWithMemberInfo` for `TypingUser` values derived from `typingUsers` and `typingMemberInfos` [#4241](https://github.com/GetStream/stream-chat-swift/pull/4241)
 - Add mention parameters to `ChatMessageController.editMessage` and `Chat.updateMessage` [#4268](https://github.com/GetStream/stream-chat-swift/pull/4268)
+- Add `memberLimit` to `ThreadQuery` for specifying the number of returned members in the thread's channel response [#4256](https://github.com/GetStream/stream-chat-swift/pull/4256)
 ### 🐞 Fixed
 - Fix mentions not being persisted when editing a message [#4268](https://github.com/GetStream/stream-chat-swift/pull/4268)
 - Fix rare crashes in `DataController.state` [#4248](https://github.com/GetStream/stream-chat-swift/pull/4248)